+++ packages/widgetbook/lib/src/core/navigation/navigation.dart
@@ -1,7 +1,6 @@
 export 'category_tree_tile.dart';
 export 'folder_tree_tile.dart';
 export 'navigation_panel.dart';
-export 'navigation_tree_node.dart';
 export 'search_field.dart';
 export 'tree.dart';
 export 'tree_node.dart';